// frontend/src/app/models/empleado.ts
export interface Empleado {
  idempleado: string;
  nombre: string;
  direccion: string;
  telefono: string;
  nit: string;
  correo: string;
  dpi: string;
  finicio: string;
  ffin: string;
  observaciones: string;
  estatus: string;
  adiciono?: string;
  fecha_adiciono?: string;
  modifico?: string;
  fecha_modificacion?: string;
  rol_id?: number;
}